requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
python-dotenv>=1.0
httpx>=0.27
sse-starlette>=2.0

# Ingestao automatica SISDPU
playwright>=1.40
pymupdf>=1.23
pytesseract>=0.3.10
Pillow>=10.0
